Force GM garbage collection every 250ms
In high throughput scenarios, e.g. basic.reject or basic.nack, messages
which belong to a mirrored queue and replicated in a GM group are
promoted to the old heap quickly. This means that garbage collection
happens only when the VM is under memory pressure, which might be too
late. When the system is under pressure, garbage collection slows it
down even further, to the point that RabbitMQ nodes can run out of
memory and crash. We want the GM process to garbage collect binaries a
lot more often, i.e. every 250ms. There is precedence for explicit GC in
the variable queue, which we do for the same reason.

Initially, we wanted to use the number of messages as the trigger for
garbage collection, but we soon discovered that different workloads
(e.g. small vs large messages) would result in unpredictable and
sub-optimal garbage collections.

Before setting fullsweep_after to 0, memory usage was 2x higher (400MB
vs 200MB) and throughput was 0.1x lower (45k vs 50k). With this setting,
there is a single heap generation resulting in a quicker GC.

The RabbitMQ deployment that this fix was tested against is:
* AWS, c4.2xlarge
* 3 nodes of RabbitMQ 3.6.11 with OTP 20.0.1
* 3 durable & auto-delete queues with 3 replicas each
* each queue master was defined on a different RabbitMQ node
* 1 consumer per queue with QOS 100
* 100 durable messages @ 1KiB each
* basic.reject operations

| Node | Message throughput | Memory usage |
|------|--------------------|--------------|
| rmq0 | 20K msg/s          | 700 MB       |
| rmq1 | 14K msg/s          | 900 MB       |
| rmq2 | 15K msg/s          | 800 MB       |

[#148892851]

Signed-off-by: Gerhard Lazu <gerhard@rabbitmq.com>